refactor: 참가자 패키지 분리
