+++ src/main/java/blackjack/model/participant/Participant.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.participant;
 
 import blackjack.model.card.Card;
 import blackjack.model.card.CardValue;
