+++ src/test/java/blackjack/model/MatchResultTest.java
@@ -9,6 +9,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import blackjack.model.card.CardValue;
 import blackjack.model.card.Deck;
 import blackjack.model.card.Suit;
+import blackjack.model.participant.Dealer;
+import blackjack.model.participant.Name;
+import blackjack.model.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
